Add files via upload
Data science and machine learning are very much applicable in such huge and various fields. One of them is retail business. Imagine if you are a retail business owner who owns a retail shop that sells hundreds of items. In a single month, there are more than a hundred transactions occurring in your shop, for instance. Each transaction is usually made with more than an item to be bought. It means there are usually more than an item in a single transaction.
In the end of the month, you would like to increase the sales in your shop, so you are trying to figure out which ways is the best to create promotions or to apply discounts on particular items. You start asking your team a questions, which products should be discounted? How do we apply the discounts? Are there any patterns that occur to our sales, so that we could apply the promotion more accurately?
Well, in order to answer that, we should try a well-known techniques in data science and machine learning called Market Basket Analysis. and here i have used  Apriori Algorithm.